requirements.txt
python-dotenv==1.0.1
boto3==1.35.2
textblob